Add endpoint to refresh an active session
As of now, we also clear any stored sessions for a username on login. This means we shouldn't have to worry about when to remove stale sessions from the database, since on login is really the only time we have any information to actually find a record to delete.

This also fixes a goof I made in getting a session. It helps to use the correct variables.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>