Implement having protected users that cannot be removed from API requests
Protected users can only be added and removed via the CLI interface. This is to ensure that in the event there is a malicious user, they cannot remove all other users and thus leave the Webby instance is a broken state without advanced intervention.

To do this necessitated refactoring the commands package for better reuseability. The database connection info is now retrieved from environment variables instead of being passed to each command.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>